Separate styles & logic in Auth
